+++ ionoscloud/data_source_ipblock.go
@@ -95,17 +95,18 @@ func dataSourceIpBlock() *schema.Resource {
 
 }
 
-func datasourceIpBlockRead(ctx context.Context, data *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	id, idOk := data.GetOk("id")
+//nolint:gocyclo
+func datasourceIpBlockRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	id, idOk := d.GetOk("id")
 
 	var name, location string
 
-	t, nameOk := data.GetOk("name")
+	t, nameOk := d.GetOk("name")
 	if nameOk {
 		name = t.(string)
 	}
 
-	t, locationOk := data.GetOk("location")
+	t, locationOk := d.GetOk("location")
 	if locationOk {
 		location = t.(string)
 	}
@@ -119,23 +120,23 @@ func datasourceIpBlockRead(ctx context.Context, data *schema.ResourceData, meta 
 	}
 
 	if !idOk && !nameOk && !locationOk {
-		return diagutil.ToDiags(data, fmt.Errorf("either id, location or name must be set"), nil)
+		return diagutil.ToDiags(d, fmt.Errorf("either id, location or name must be set"), nil)
 	}
 	if idOk {
 		ipBlock, apiResponse, err = client.IPBlocksApi.IpblocksFindById(ctx, id.(string)).Execute()
 		logApiRequestTime(apiResponse)
 		if err != nil {
-			return diagutil.ToDiags(data, fmt.Errorf("error getting ip block with id %s %w", id.(string), err), &diagutil.ErrorContext{StatusCode: apiResponse.StatusCode})
+			return diagutil.ToDiags(d, fmt.Errorf("error getting ip block with id %s %w", id.(string), err), &diagutil.ErrorContext{StatusCode: apiResponse.StatusCode})
 		}
 		if nameOk {
 			if ipBlock.Properties != nil && *ipBlock.Properties.Name != name {
-				return diagutil.ToDiags(data, fmt.Errorf("name of ip block (UUID=%s, name=%s) does not match expected name: %s",
+				return diagutil.ToDiags(d, fmt.Errorf("name of ip block (UUID=%s, name=%s) does not match expected name: %s",
 					*ipBlock.Id, *ipBlock.Properties.Name, name), nil)
 			}
 		}
 		if locationOk {
 			if ipBlock.Properties != nil && *ipBlock.Properties.Location != location {
-				return diagutil.ToDiags(data, fmt.Errorf("location of ip block (UUID=%s, location=%s) does not match expected location: %s",
+				return diagutil.ToDiags(d, fmt.Errorf("location of ip block (UUID=%s, location=%s) does not match expected location: %s",
 					*ipBlock.Id, *ipBlock.Properties.Location, location), nil)
 			}
 		}
@@ -146,7 +147,7 @@ func datasourceIpBlockRead(ctx context.Context, data *schema.ResourceData, meta 
 		logApiRequestTime(apiResponse)
 
 		if err != nil {
-			return diagutil.ToDiags(data, fmt.Errorf("an error occurred while fetching ipBlocks: %w ", err), &diagutil.ErrorContext{StatusCode: apiResponse.StatusCode})
+			return diagutil.ToDiags(d, fmt.Errorf("an error occurred while fetching ipBlocks: %w ", err), &diagutil.ErrorContext{StatusCode: apiResponse.StatusCode})
 		}
 
 		var results []ionoscloud.IpBlock
@@ -159,7 +160,7 @@ func datasourceIpBlockRead(ctx context.Context, data *schema.ResourceData, meta 
 			}
 
 			if results == nil {
-				return diagutil.ToDiags(data, fmt.Errorf("no ip block found with the specified criteria name %s", name), nil)
+				return diagutil.ToDiags(d, fmt.Errorf("no ip block found with the specified criteria name %s", name), nil)
 			}
 		}
 
@@ -183,17 +184,17 @@ func datasourceIpBlockRead(ctx context.Context, data *schema.ResourceData, meta 
 		}
 
 		if results == nil || len(results) == 0 {
-			return diagutil.ToDiags(data, fmt.Errorf("no ip block found with the specified criteria name = %s, location = %s", name, location), nil)
+			return diagutil.ToDiags(d, fmt.Errorf("no ip block found with the specified criteria name = %s, location = %s", name, location), nil)
 		} else if len(results) > 1 {
-			return diagutil.ToDiags(data, fmt.Errorf("more than one ip block found with the specified criteria name = %s, location = %s", name, location), nil)
+			return diagutil.ToDiags(d, fmt.Errorf("more than one ip block found with the specified criteria name = %s, location = %s", name, location), nil)
 		} else {
 			ipBlock = results[0]
 		}
 
 	}
 
-	if err := IpBlockSetData(data, &ipBlock); err != nil {
-		return diagutil.ToDiags(data, err, nil)
+	if err := IpBlockSetData(d, &ipBlock); err != nil {
+		return diagutil.ToDiags(d, err, nil)
 	}
 
 	return nil